+++ tests/parse_args.php
@@ -35,7 +35,7 @@ class argParser
             'HTTPSSERVER' => 'localhost',
             'HTTPSURI' => null,
             // example alternative:
-            //'HTTPSSERVER' => 'gggeek.altervista.org',
+            //'HTTPSSERVER' => 'tanoconsulting.com',
             //'HTTPSURI' => '/sw/xmlrpc/demo/server/server.php',
             'HTTPSIGNOREPEER' => false,
             'HTTPSVERIFYHOST' => 2,